+++ packages/assets-controllers/jest.config.js
@@ -17,10 +17,10 @@ module.exports = merge(baseConfig, {
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
-      branches: 88.36,
-      functions: 97.08,
-      lines: 97.23,
-      statements: 97.28,
+      branches: 88.93,
+      functions: 96.71,
+      lines: 97.34,
+      statements: 97.4,
     },
   },