+++ packages/assets-controllers/src/index.ts
@@ -19,9 +19,35 @@ export type {
   TokenDetectionControllerStateChangeEvent,
 } from './TokenDetectionController';
 export { TokenDetectionController } from './TokenDetectionController';
-export * from './TokenListController';
-export * from './TokenRatesController';
-export * from './TokensController';
+export type {
+  TokenListState,
+  TokenListToken,
+  TokenListMap,
+  TokenListStateChange,
+  TokenListControllerEvents,
+  GetTokenListState,
+  TokenListControllerActions,
+  TokenListControllerMessenger,
+} from './TokenListController';
+export { TokenListController } from './TokenListController';
+export type {
+  Token,
+  TokenRatesConfig,
+  ContractExchangeRates,
+  TokenRatesState,
+} from './TokenRatesController';
+export { TokenRatesController } from './TokenRatesController';
+export type {
+  TokensConfig,
+  TokensState,
+  TokensControllerActions,
+  TokensControllerGetStateAction,
+  TokensControllerAddDetectedTokensAction,
+  TokensControllerEvents,
+  TokensControllerStateChangeEvent,
+  TokensControllerMessenger,
+} from './TokensController';
+export { TokensController } from './TokensController';
 export {
   isTokenDetectionSupportedForNetwork,
   formatIconUrlWithProxy,