+++ packages/transaction-controller/src/helpers/PendingTransactionTracker.ts
@@ -144,6 +144,24 @@ export class PendingTransactionTracker {
     });
   }
 
+  /**
+   * Force checks the network if the given transaction is confirmed and updates it's status.
+   *
+   * @param txMeta - The transaction to check
+   */
+  async forceCheckTransaction(txMeta: TransactionMeta) {
+    const nonceGlobalLock = await this.#nonceTracker.getGlobalLock();
+
+    try {
+      await this.#checkTransaction(txMeta);
+    } catch (error) {
+      /* istanbul ignore next */
+      log('Failed to check transaction', error);
+    } finally {
+      nonceGlobalLock.releaseLock();
+    }
+  }
+
   #start() {
     if (this.#running) {
       return;